+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 ### Fixed
 
 - visualization error if no lesion data is available (Robert Arnold)
-- ambiguity for annotation times 0 and NaN (Robert Arnold) 
+- ambiguity for annotation times 0 and NaN in CartoPoint (Robert Arnold)
+- export error if no VisiTag sites were loaded/found (Robert Arnold)
+- load error if no lesion data in .pyceps file (Robert Arnold)
+- import error for single channel reference annotation config (algorithm = 0) (Robert Arnold)
 
 ### Removed